The shrapnel effect was still slowing down the game increasingly the more robots had been destroyed. If there were a great lot of light emitting objects, the player ship was spinning out of control or the program crashed. No shrapnel effects will be created during application of the 'spaniard' cheat.
git-svn-id: svn://svn.code.sf.net/p/d2x-xl/code/trunk@628 01821946-851b-0410-b7f4-f1477c70e57a

+++ main/playsave.c
@@ -144,7 +144,7 @@ for (pi = pszTag; *pszIdent; h++) {
 		strcpy (pi, pszIdent);
 		break;
 		}
-	memcpy (pi, pszIdent, l = ph - pszIdent);
+	memcpy (pi, pszIdent, l = (int) (ph - pszIdent));
 	sprintf (pi + l, "[%d]", h ? j : i);
 	pi = pszTag + strlen (pszTag);
 	strcat (pszTag, pszIdent = strchr (ph + 1, ']') + 1);
@@ -160,7 +160,7 @@ int RegisterParam (void *valP, char *pszIdent, int i, int j, ubyte nSize)
 	int		l;
 	tParam	*pp;
 
-l = strlen (MakeTag (szTag, pszIdent, i, j));
+l = (int) strlen (MakeTag (szTag, pszIdent, i, j));
 pp = (tParam *) D2_ALLOC (sizeof (tParam) + l);
 if (!pp)
 	return 0;
@@ -576,8 +576,8 @@ done:
 
 strcat (szVal, "\n");
 #endif
-CFWrite (pp->szTag, 1, strlen (pp->szTag), fp);
-CFWrite (szVal, 1, strlen (szVal), fp);
+CFWrite (pp->szTag, 1, (int) strlen (pp->szTag), fp);
+CFWrite (szVal, 1, (int) strlen (szVal), fp);
 fflush (fp->file);
 return 1;
 }
